Move CakeErrorStringifier to separate file
